+++ src/classes/Dictionary.h
@@ -12,7 +12,7 @@
 using namespace std;
 
 struct cacheElem{
-	unsigned int tag;
+	unsigned long long int tag;
 	bool valid;
 	bool dirty;
 };
@@ -26,7 +26,7 @@ public:
 	//checks to see if a certain index contains the tag (valid)
 	bool check(unsigned int index, unsigned long long int tag);
 	//change the specified index to the new tag
-	void update(unsigned int index, unsigned int tag);
+	void update(unsigned int index, unsigned long long int tag);
 	//flush the cache
 	void flush(void);
 	//get the element at a specific location
